Renaming pitch_bend_change file to pitch_bend (part II of previous commit). #3
