+++ examples/midi_inoutdemo.py
@@ -12,7 +12,7 @@ from adafruit_midi.channel_pressure        import ChannelPressure
 from adafruit_midi.control_change          import ControlChange
 from adafruit_midi.note_off                import NoteOff
 from adafruit_midi.note_on                 import NoteOn
-from adafruit_midi.pitch_bend_change       import PitchBend
+from adafruit_midi.pitch_bend       import PitchBend
 from adafruit_midi.polyphonic_key_pressure import PolyphonicKeyPressure
 from adafruit_midi.program_change          import ProgramChange
 from adafruit_midi.start                   import Start
