+++ examples/midi_simpletest2.py
@@ -5,7 +5,7 @@ import adafruit_midi
 from adafruit_midi.control_change          import ControlChange
 from adafruit_midi.note_off                import NoteOff
 from adafruit_midi.note_on                 import NoteOn
-from adafruit_midi.pitch_bend_change       import PitchBend
+from adafruit_midi.pitch_bend       import PitchBend
 
 midi = adafruit_midi.MIDI(out_channel=0)
